Update urls and archive.sh to handle vimeo and channel9 urls.
Created a fixup.sh script to migrate the old filenames to the less
verbose now used.

+++ archive.sh
@@ -4,9 +4,15 @@ export script=$(basename "$0")
 export dir=$(cd "$(dirname "$0")"; pwd)
 export iam=${dir}/${script}
 
-for file in youtube*; do
-  subdir=$(echo "${file}" | awk -F- '{print $3}' | sort | uniq)
-  [ ! -d "${subdir}" ] && mkdir "${subdir}"
-  echo mv "${file}" "${subdir}"
-  mv "${file}" "${subdir}"
+for file in youtube-* vimeo-* channel9-*; do
+  service=$(echo "${file}" | awk -F- '{print $1}' | sort | uniq)
+  user=$(echo "${file}" | awk -F- '{print $3}')
+  subdir="${service}/"$(echo ${user} | sort | uniq)
+  [ ! -d "${subdir}" ] && mkdir -p "${subdir}"
+
+  if [ -e "${file}" ]; then
+      new_name=$(echo ${file} | sed -e "s|${service}-||" -e "s|${user}-||")
+      echo mv "${file}" "${subdir}/${new_name}"
+      mv "${file}" "${subdir}"
+  fi
 done
